refactor: 추천 여행지 목록 UI 기능 컴포넌트로 분리 및 구조 리팩토링
